requirements.txt
click==8.1.7
pyyaml==6.0.1
pytest==7.4.4